Move BusMessage to common
