+++ nautilus_core/common/src/python/mod.rs
@@ -25,14 +25,14 @@ pub mod msgbus;
 pub mod timer;
 pub mod versioning;
 
-use handler::PythonMessageHandler;
 use pyo3::prelude::*;
 
 /// Loaded as nautilus_pyo3.common
 #[pymodule]
 pub fn common(_: Python<'_>, m: &PyModule) -> PyResult<()> {
     m.add_class::<crate::msgbus::MessageBus>()?;
-    m.add_class::<PythonMessageHandler>()?;
+    m.add_class::<crate::msgbus::database::BusMessage>()?;
+    m.add_class::<crate::python::handler::PythonMessageHandler>()?;
     m.add_class::<crate::enums::ComponentState>()?;
     m.add_class::<crate::enums::ComponentTrigger>()?;
     m.add_class::<crate::enums::LogColor>()?;
